Add Blank to type system. Nullable struct is not suitable, because blank is a type of value that is being passed through formula evaluation.
